The Project seems to be at a point we can upgrade. The basiucs woirk as per the live version published in 2018.

+++ Pages/ReoccuringOrderDetails.aspx.cs
@@ -65,7 +65,7 @@ namespace TrackerDotNet.Pages
             }
         }
 
-        private void PutDataFromForm(long pReoccuringOrderID)
+        private void PutDataFromForm(int pReoccuringOrderID)
         {
             ReoccuringOrderTbl reoccuringOrderById = new ReoccuringOrderDAL().GetByReoccuringOrderByID(pReoccuringOrderID);
             if (reoccuringOrderById == null)
